+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2023-11-03
+
 ## Added
 
 - `NewHandler` to be used as a middleware with `WithContext` to add logger to the context.
@@ -35,7 +37,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First public release.
 
-[unreleased]: https://gitlab.com/tozd/go/zerolog/-/compare/v0.3.0...main
+[unreleased]: https://gitlab.com/tozd/go/zerolog/-/compare/v0.4.0...main
+[0.4.0]: https://gitlab.com/tozd/go/zerolog/-/compare/v0.3.0...v0.4.0
 [0.3.0]: https://gitlab.com/tozd/go/zerolog/-/compare/v0.2.0...v0.3.0
 [0.2.0]: https://gitlab.com/tozd/go/zerolog/-/compare/v0.1.0...v0.2.0
 [0.1.0]: https://gitlab.com/tozd/go/zerolog/-/tags/v0.1.0